Move unit test to TestAutoInference and fix EntryPointCatalog test.
