+++ ZBaselines/Common/EntryPoints/core_ep-list.tsv
@@ -15,11 +15,13 @@ Models.NaiveCalibrator	Apply a Naive calibrator to an input model	Microsoft.ML.R
 Models.OneVersusAll	One-vs-All macro (OVA)	Microsoft.ML.Runtime.EntryPoints.OneVersusAllMacro	OVA	Microsoft.ML.Runtime.EntryPoints.OneVersusAllMacro+Arguments	Microsoft.ML.Runtime.EntryPoints.CommonOutputs+MacroOutput`1[Microsoft.ML.Runtime.EntryPoints.OneVersusAllMacro+Output]
 Models.OvaModelCombiner	Combines a sequence of PredictorModels into a single model	Microsoft.ML.Runtime.Learners.OvaPredictor	CombineOvaModels	Microsoft.ML.Runtime.EntryPoints.ModelOperations+CombineOvaPredictorModelsInput	Microsoft.ML.Runtime.EntryPoints.ModelOperations+PredictorModelOutput
 Models.PAVCalibrator	Apply a PAV calibrator to an input model	Microsoft.ML.Runtime.Internal.Calibration.Calibrate	Pav	Microsoft.ML.Runtime.Internal.Calibration.Calibrate+NoArgumentsInput	Microsoft.ML.Runtime.EntryPoints.CommonOutputs+CalibratorOutput
+Models.PipelineSweeper	AutoML pipeline sweeping optimzation macro.	Microsoft.ML.Runtime.EntryPoints.PipelineSweeperMacro	PipelineSweep	Microsoft.ML.Runtime.EntryPoints.PipelineSweeperMacro+Arguments	Microsoft.ML.Runtime.EntryPoints.CommonOutputs+MacroOutput`1[Microsoft.ML.Runtime.EntryPoints.PipelineSweeperMacro+Output]
 Models.PlattCalibrator	Apply a Platt calibrator to an input model	Microsoft.ML.Runtime.Internal.Calibration.Calibrate	Platt	Microsoft.ML.Runtime.Internal.Calibration.Calibrate+NoArgumentsInput	Microsoft.ML.Runtime.EntryPoints.CommonOutputs+CalibratorOutput
 Models.QuantileRegressionEvaluator	Evaluates a quantile regression scored dataset.	Microsoft.ML.Runtime.Data.Evaluate	QuantileRegression	Microsoft.ML.Runtime.Data.QuantileRegressionMamlEvaluator+Arguments	Microsoft.ML.Runtime.EntryPoints.CommonOutputs+CommonEvaluateOutput
 Models.RankerEvaluator	Evaluates a ranking scored dataset.	Microsoft.ML.Runtime.Data.Evaluate	Ranking	Microsoft.ML.Runtime.Data.RankerMamlEvaluator+Arguments	Microsoft.ML.Runtime.EntryPoints.CommonOutputs+CommonEvaluateOutput
 Models.RegressionEvaluator	Evaluates a regression scored dataset.	Microsoft.ML.Runtime.Data.Evaluate	Regression	Microsoft.ML.Runtime.Data.RegressionMamlEvaluator+Arguments	Microsoft.ML.Runtime.EntryPoints.CommonOutputs+CommonEvaluateOutput
 Models.Summarizer	Summarize a linear regression predictor.	Microsoft.ML.Runtime.EntryPoints.SummarizePredictor	Summarize	Microsoft.ML.Runtime.EntryPoints.SummarizePredictor+Input	Microsoft.ML.Runtime.EntryPoints.CommonOutputs+SummaryOutput
+Models.SweepResultExtractor	Extracts the sweep result.	Microsoft.ML.Runtime.EntryPoints.PipelineSweeperMacro	ExtractSweepResult	Microsoft.ML.Runtime.EntryPoints.PipelineSweeperMacro+ResultInput	Microsoft.ML.Runtime.EntryPoints.PipelineSweeperMacro+Output
 Models.TrainTestBinaryEvaluator	Train test for binary classification	Microsoft.ML.Runtime.EntryPoints.TrainTestBinaryMacro	TrainTestBinary	Microsoft.ML.Runtime.EntryPoints.TrainTestBinaryMacro+Arguments	Microsoft.ML.Runtime.EntryPoints.CommonOutputs+MacroOutput`1[Microsoft.ML.Runtime.EntryPoints.TrainTestBinaryMacro+Output]
 Models.TrainTestEvaluator	General train test for any supported evaluator	Microsoft.ML.Runtime.EntryPoints.TrainTestMacro	TrainTest	Microsoft.ML.Runtime.EntryPoints.TrainTestMacro+Arguments	Microsoft.ML.Runtime.EntryPoints.CommonOutputs+MacroOutput`1[Microsoft.ML.Runtime.EntryPoints.TrainTestMacro+Output]
 Trainers.AveragedPerceptronBinaryClassifier	Train a Average perceptron.	Microsoft.ML.Runtime.Learners.AveragedPerceptronTrainer	TrainBinary	Microsoft.ML.Runtime.Learners.AveragedPerceptronTrainer+Arguments	Microsoft.ML.Runtime.EntryPoints.CommonOutputs+BinaryClassificationOutput
